refactor: extract rule element operation tests into dedicated files
