+++ tests/unit/rule-elements/pf2e-visioner-effect.test.js
@@ -1,6 +1,5 @@
 import { beforeEach, describe, expect, it, jest } from '@jest/globals';
 import { SourceTracker } from '../../../scripts/rule-elements/SourceTracker.js';
-import { ActionQualifier } from '../../../scripts/rule-elements/operations/ActionQualifier.js';
 
 describe('SourceTracker', () => {
   let mockToken;
@@ -215,431 +214,17 @@ describe('SourceTracker', () => {
   });
 });
 
-describe('DetectionModeModifier', () => {
-  let mockToken, mockDetectionsModes;
 
-  beforeEach(() => {
-    mockDetectionsModes = [
-      { id: 'visual', enabled: true },
-      { id: 'imprecise-hearing', enabled: true },
-    ];
-    mockToken = {
-      id: 'test-token',
-      document: {
-        id: 'test-token',
-        detectionModes: mockDetectionsModes,
-        getFlag: jest.fn(() => ({})),
-        setFlag: jest.fn(() => Promise.resolve()),
-        update: jest.fn(() => Promise.resolve()),
-      },
-    };
-  });
 
-  describe('applyDetectionModeModifications', () => {
-    it('should save original detection modes before modifying', async () => {
-      const { DetectionModeModifier } = await import(
-        '../../../scripts/rule-elements/operations/DetectionModeModifier.js'
-      );
 
-      await DetectionModeModifier.applyDetectionModeModifications(
-        mockToken,
-        { visual: { enabled: false } },
-        'test-rule-element',
-        null,
-      );
 
-      expect(mockToken.document.getFlag).toHaveBeenCalled();
-      expect(mockToken.document.update).toHaveBeenCalled();
-    });
-  });
-});
 
-describe('DistanceBasedVisibility', () => {
-  let mockToken;
 
-  beforeEach(() => {
-    mockToken = {
-      id: 'test-token',
-      document: {
-        id: 'test-token',
-        getFlag: jest.fn(() => null),
-        setFlag: jest.fn(() => Promise.resolve()),
-        unsetFlag: jest.fn(() => Promise.resolve()),
-      },
-    };
-  });
 
-  describe('applyDistanceBasedVisibility', () => {
-    it('should set distance-based visibility flag', async () => {
-      const { DistanceBasedVisibility } = await import(
-        '../../../scripts/rule-elements/operations/DistanceBasedVisibility.js'
-      );
 
-      const operation = {
-        distanceBands: [
-          { minDistance: 0, maxDistance: 20, state: 'observed' },
-          { minDistance: 20, maxDistance: null, state: 'concealed' },
-        ],
-        direction: 'to',
-        observers: 'all',
-        source: 'test-source',
-      };
 
-      await DistanceBasedVisibility.applyDistanceBasedVisibility(operation, mockToken);
 
-      expect(mockToken.document.setFlag).toHaveBeenCalledWith(
-        'pf2e-visioner',
-        'distanceBasedVisibility',
-        expect.objectContaining({
-          active: true,
-          source: 'test-source',
-          distanceBands: operation.distanceBands,
-        }),
-      );
-    });
 
-    it('should return correct distance band for distance', () => {
-      const {
-        DistanceBasedVisibility,
-      } = require('../../../scripts/rule-elements/operations/DistanceBasedVisibility.js');
-
-      const distanceBands = [
-        { minDistance: 0, maxDistance: 20, state: 'observed' },
-        { minDistance: 20, maxDistance: 40, state: 'concealed' },
-        { minDistance: 40, maxDistance: null, state: 'hidden' },
-      ];
-
-      expect(DistanceBasedVisibility.getApplicableDistanceBand(10, distanceBands).state).toBe(
-        'observed',
-      );
-      expect(DistanceBasedVisibility.getApplicableDistanceBand(30, distanceBands).state).toBe(
-        'concealed',
-      );
-      expect(DistanceBasedVisibility.getApplicableDistanceBand(50, distanceBands).state).toBe(
-        'hidden',
-      );
-    });
-
-    it('should remove distance-based visibility on removal', async () => {
-      const { DistanceBasedVisibility } = await import(
-        '../../../scripts/rule-elements/operations/DistanceBasedVisibility.js'
-      );
-
-      const operation = { source: 'test-source' };
-
-      await DistanceBasedVisibility.removeDistanceBasedVisibility(operation, mockToken);
-
-      expect(mockToken.document.unsetFlag).toHaveBeenCalledWith(
-        'pf2e-visioner',
-        'distanceBasedVisibility',
-      );
-    });
-  });
-});
-
-describe('LightingModifier', () => {
-  let mockToken;
-
-  beforeEach(() => {
-    mockToken = {
-      id: 'test-token',
-      document: {
-        id: 'test-token',
-        setFlag: jest.fn(() => Promise.resolve()),
-      },
-    };
-  });
-
-  describe('applyLightingModification', () => {
-    it('should set lighting modification flag', async () => {
-      const { LightingModifier } = await import(
-        '../../../scripts/rule-elements/operations/LightingModifier.js'
-      );
-
-      const operation = {
-        lightingLevel: 'bright',
-        source: 'test-light',
-        priority: 100,
-      };
-
-      await LightingModifier.applyLightingModification(operation, mockToken);
-
-      expect(mockToken.document.setFlag).toHaveBeenCalledWith(
-        'pf2e-visioner',
-        'lightingModification.test-light',
-        expect.objectContaining({
-          lightingLevel: 'bright',
-          type: 'test-light',
-          priority: 100,
-        }),
-      );
-    });
-  });
-});
-
-describe('OffGuardSuppression', () => {
-  let mockToken;
-
-  beforeEach(() => {
-    mockToken = {
-      id: 'test-token',
-      document: {
-        id: 'test-token',
-        getFlag: jest.fn(() => null),
-        setFlag: jest.fn(() => Promise.resolve()),
-        unsetFlag: jest.fn(() => Promise.resolve()),
-      },
-    };
-  });
-
-  describe('applyOffGuardSuppression', () => {
-    it('should set off guard suppression flag', async () => {
-      const { OffGuardSuppression } = await import(
-        '../../../scripts/rule-elements/operations/OffGuardSuppression.js'
-      );
-
-      const operation = {
-        suppressedStates: ['hidden', 'undetected'],
-        source: 'test-suppression',
-      };
-
-      await OffGuardSuppression.applyOffGuardSuppression(operation, mockToken);
-
-      expect(mockToken.document.setFlag).toHaveBeenCalledWith(
-        'pf2e-visioner',
-        expect.stringContaining('offGuardSuppression.'),
-        expect.objectContaining({
-          suppressedStates: ['hidden', 'undetected'],
-        }),
-      );
-    });
-
-    it('should check if state should suppress off guard', async () => {
-      const { OffGuardSuppression } = await import(
-        '../../../scripts/rule-elements/operations/OffGuardSuppression.js'
-      );
-
-      mockToken.document.getFlag.mockReturnValue({
-        'test-suppression': {
-          suppressedStates: ['hidden', 'undetected'],
-        },
-      });
-
-      const shouldSuppress = OffGuardSuppression.shouldSuppressOffGuardForState(
-        mockToken,
-        'hidden',
-      );
-      expect(shouldSuppress).toBe(true);
-
-      const shouldNotSuppress = OffGuardSuppression.shouldSuppressOffGuardForState(
-        mockToken,
-        'concealed',
-      );
-      expect(shouldNotSuppress).toBe(false);
-    });
-
-    it('should remove off guard suppression on removal', async () => {
-      const { OffGuardSuppression } = await import(
-        '../../../scripts/rule-elements/operations/OffGuardSuppression.js'
-      );
-
-      const operation = { source: 'test-suppression' };
-      mockToken.document.getFlag.mockReturnValue({
-        'test-suppression': {
-          suppressedStates: ['hidden', 'undetected'],
-        },
-      });
-
-      await OffGuardSuppression.removeOffGuardSuppression(operation, mockToken);
-
-      expect(mockToken.document.unsetFlag).toHaveBeenCalledWith(
-        'pf2e-visioner',
-        'offGuardSuppression.test-suppression',
-      );
-    });
-  });
-});
-
-describe('SenseModifier', () => {
-  let mockToken;
-  let mockActor;
-
-  beforeEach(() => {
-    mockActor = {
-      system: {
-        perception: {
-          senses: [
-            { type: 'vision', acuity: 'precise', range: Infinity },
-            { type: 'hearing', acuity: 'imprecise', range: Infinity },
-          ],
-        },
-      },
-    };
-
-    mockToken = {
-      actor: {
-        ...mockActor,
-        update: jest.fn(() => Promise.resolve()),
-      },
-      document: {
-        getFlag: jest.fn(() => ({})),
-        setFlag: jest.fn(() => Promise.resolve()),
-        update: jest.fn(() => Promise.resolve()),
-      },
-    };
-  });
-
-  it('should save original senses before modifying', async () => {
-    const { SenseModifier } = await import(
-      '../../../scripts/rule-elements/operations/SenseModifier.js'
-    );
-
-    await SenseModifier.applySenseModifications(
-      mockToken,
-      { vision: { enabled: false } },
-      'test-rule-element',
-      null,
-    );
-
-    expect(mockToken.document.getFlag).toHaveBeenCalled();
-    expect(mockToken.actor.update).toHaveBeenCalled();
-  });
-});
-
-describe('ActionQualifier', () => {
-  let mockToken;
-
-  beforeEach(() => {
-    mockToken = {
-      document: {
-        getFlag: jest.fn(() => ({})),
-        setFlag: jest.fn(() => Promise.resolve()),
-      },
-    };
-  });
-
-  describe('applyActionQualifications', () => {
-    it('should set action qualifications on token', async () => {
-      const operation = {
-        source: 'blur-spell',
-        qualifications: {
-          hide: { canUseThisConcealment: false },
-          sneak: { endPositionQualifies: false },
-        },
-      };
-
-      await ActionQualifier.applyActionQualifications(operation, mockToken);
-
-      expect(mockToken.document.setFlag).toHaveBeenCalledWith(
-        'pf2e-visioner',
-        'actionQualifications.blur-spell',
-        expect.objectContaining({
-          id: 'blur-spell',
-          qualifications: operation.qualifications,
-        }),
-      );
-    });
-  });
-
-  describe('getActionQualifications', () => {
-    it('should return qualifications for specific action', () => {
-      mockToken.document.getFlag.mockReturnValue({
-        'blur-spell': {
-          id: 'blur-spell',
-          priority: 100,
-          qualifications: {
-            hide: { canUseThisConcealment: false },
-            sneak: { endPositionQualifies: false },
-          },
-        },
-      });
-
-      const qualifications = ActionQualifier.getActionQualifications(mockToken, 'hide');
-
-      expect(qualifications).toHaveLength(1);
-      expect(qualifications[0].canUseThisConcealment).toBe(false);
-    });
-  });
-
-  describe('canUseConcealment', () => {
-    it('should return false when qualification disallows concealment', () => {
-      mockToken.document.getFlag.mockReturnValue({
-        'blur-spell': {
-          id: 'blur-spell',
-          priority: 100,
-          qualifications: {
-            hide: { qualifiesOnConcealment: false },
-          },
-        },
-      });
-
-      const canUse = ActionQualifier.canUseConcealment(mockToken, 'hide');
-      expect(canUse).toBe(false);
-    });
-
-    it('should return true when no qualifications exist', () => {
-      const canUse = ActionQualifier.canUseConcealment(mockToken, 'hide');
-      expect(canUse).toBe(true);
-    });
-  });
-
-  describe('endPositionQualifies', () => {
-    it('should return false when qualification disallows end position', () => {
-      mockToken.document.getFlag.mockReturnValue({
-        'blur-spell': {
-          id: 'blur-spell',
-          priority: 100,
-          qualifications: {
-            sneak: { endPositionQualifies: false },
-          },
-        },
-      });
-
-      const qualifies = ActionQualifier.endPositionQualifies(mockToken, 'sneak');
-      expect(qualifies).toBe(false);
-    });
-
-    it('should return true when no qualifications exist', () => {
-      const qualifies = ActionQualifier.endPositionQualifies(mockToken, 'sneak');
-      expect(qualifies).toBe(true);
-    });
-  });
-
-  describe('checkHidePrerequisites', () => {
-    it('should return cannot hide when no qualifying sources', () => {
-      mockToken.document.getFlag.mockReturnValue({
-        'blur-spell': {
-          id: 'blur-spell',
-          qualifications: {
-            hide: { canUseThisConcealment: false },
-          },
-        },
-      });
-
-      const result = ActionQualifier.checkHidePrerequisites(mockToken);
-
-      expect(result.canHide).toBe(false);
-      expect(result.qualifyingConcealment).toBe(0);
-    });
-  });
-
-  describe('checkSneakPrerequisites', () => {
-    it('should return does not qualify when end position disqualified', () => {
-      mockToken.document.getFlag.mockReturnValue({
-        'blur-spell': {
-          id: 'blur-spell',
-          qualifications: {
-            sneak: { endPositionQualifies: false },
-          },
-        },
-      });
-
-      const result = ActionQualifier.checkSneakPrerequisites(mockToken, 'end');
-
-      expect(result.qualifies).toBe(false);
-    });
-  });
-});
 
 describe('Rule Element Integration', () => {
   it('should have all required operation types', () => {
@@ -690,250 +275,9 @@ describe('Rule Element Integration', () => {
   });
 });
 
-describe('VisibilityOverride', () => {
-  let mockSubjectToken;
-  let mockObserverTokens;
 
-  beforeEach(async () => {
-    const { VisibilityOverride } = await import(
-      '../../../scripts/rule-elements/operations/VisibilityOverride.js'
-    );
 
-    mockSubjectToken = {
-      id: 'subject-token',
-      actor: { hasPlayerOwner: true },
-      document: {
-        id: 'subject-token',
-        setFlag: jest.fn(() => Promise.resolve()),
-        unsetFlag: jest.fn(() => Promise.resolve()),
-      },
-    };
 
-    mockObserverTokens = [
-      {
-        id: 'observer-1',
-        actor: { hasPlayerOwner: true },
-        document: { id: 'observer-1' },
-      },
-      {
-        id: 'observer-2',
-        actor: { hasPlayerOwner: false, token: { disposition: -1 } },
-        document: { id: 'observer-2' },
-      },
-    ];
-
-    global.canvas = {
-      tokens: {
-        placeables: [mockSubjectToken, ...mockObserverTokens],
-      },
-      grid: {
-        measureDistance: jest.fn(() => 10),
-      },
-    };
-  });
-
-  describe('getObserverTokens', () => {
-    it('should return all tokens when observers is "all"', async () => {
-      const { VisibilityOverride } = await import(
-        '../../../scripts/rule-elements/operations/VisibilityOverride.js'
-      );
-      const tokens = VisibilityOverride.getObserverTokens(mockSubjectToken, 'all');
-
-      expect(tokens.length).toBe(2);
-      expect(tokens.map((t) => t.id)).toEqual(['observer-1', 'observer-2']);
-    });
-
-    it('should filter allies when observers is "allies"', async () => {
-      const { VisibilityOverride } = await import(
-        '../../../scripts/rule-elements/operations/VisibilityOverride.js'
-      );
-      const tokens = VisibilityOverride.getObserverTokens(mockSubjectToken, 'allies');
-
-      expect(tokens.length).toBe(1);
-      expect(tokens[0].id).toBe('observer-1');
-    });
-
-    it('should filter enemies when observers is "enemies"', async () => {
-      const { VisibilityOverride } = await import(
-        '../../../scripts/rule-elements/operations/VisibilityOverride.js'
-      );
-      const tokens = VisibilityOverride.getObserverTokens(mockSubjectToken, 'enemies');
-
-      expect(tokens.length).toBe(1);
-      expect(tokens[0].id).toBe('observer-2');
-    });
-
-    it('should filter specific tokens when observers is "specific"', async () => {
-      const { VisibilityOverride } = await import(
-        '../../../scripts/rule-elements/operations/VisibilityOverride.js'
-      );
-      const tokenIds = ['observer-2'];
-      const tokens = VisibilityOverride.getObserverTokens(
-        mockSubjectToken,
-        'specific',
-        null,
-        tokenIds,
-      );
-
-      expect(tokens.length).toBe(1);
-      expect(tokens[0].id).toBe('observer-2');
-    });
-
-    it('should apply range filter when range is specified', async () => {
-      const { VisibilityOverride } = await import(
-        '../../../scripts/rule-elements/operations/VisibilityOverride.js'
-      );
-      global.canvas.grid.measureDistance = jest.fn().mockReturnValueOnce(5).mockReturnValueOnce(25);
-
-      const tokens = VisibilityOverride.getObserverTokens(mockSubjectToken, 'all', 20);
-
-      expect(tokens.length).toBe(1);
-      expect(tokens[0].id).toBe('observer-1');
-    });
-  });
-
-  describe('areAllies', () => {
-    it('should return true for PC vs PC', async () => {
-      const { VisibilityOverride } = await import(
-        '../../../scripts/rule-elements/operations/VisibilityOverride.js'
-      );
-      const actor1 = { hasPlayerOwner: true };
-      const actor2 = { hasPlayerOwner: true };
-
-      expect(VisibilityOverride.areAllies(actor1, actor2)).toBe(true);
-    });
-
-    it('should return false for PC vs NPC', async () => {
-      const { VisibilityOverride } = await import(
-        '../../../scripts/rule-elements/operations/VisibilityOverride.js'
-      );
-      const actor1 = { hasPlayerOwner: true };
-      const actor2 = { hasPlayerOwner: false, token: { disposition: -1 } };
-
-      expect(VisibilityOverride.areAllies(actor1, actor2)).toBe(false);
-    });
-
-    it('should return true for NPCs with same disposition', async () => {
-      const { VisibilityOverride } = await import(
-        '../../../scripts/rule-elements/operations/VisibilityOverride.js'
-      );
-      const actor1 = { hasPlayerOwner: false, token: { disposition: 1 } };
-      const actor2 = { hasPlayerOwner: false, token: { disposition: 1 } };
-
-      expect(VisibilityOverride.areAllies(actor1, actor2)).toBe(true);
-    });
-  });
-});
-
-describe('CoverOverride', () => {
-  let mockSubjectToken;
-  let mockTargetTokens;
-
-  beforeEach(async () => {
-    mockSubjectToken = {
-      id: 'subject-token',
-      actor: { hasPlayerOwner: true },
-      document: {
-        id: 'subject-token',
-        setFlag: jest.fn(() => Promise.resolve()),
-        getFlag: jest.fn(() => null),
-      },
-    };
-
-    mockTargetTokens = [
-      {
-        id: 'target-1',
-        actor: { hasPlayerOwner: true },
-        document: { id: 'target-1' },
-      },
-      {
-        id: 'target-2',
-        actor: { hasPlayerOwner: false, token: { disposition: -1 } },
-        document: { id: 'target-2' },
-      },
-    ];
-
-    global.canvas = {
-      tokens: {
-        placeables: [mockSubjectToken, ...mockTargetTokens],
-      },
-      grid: {
-        measureDistance: jest.fn(() => 10),
-      },
-    };
-  });
-
-  describe('getTargetTokens', () => {
-    it('should return all tokens when targets is "all"', async () => {
-      const { CoverOverride } = await import(
-        '../../../scripts/rule-elements/operations/CoverOverride.js'
-      );
-      const tokens = CoverOverride.getTargetTokens(mockSubjectToken, 'all');
-
-      expect(tokens.length).toBe(2);
-    });
-
-    it('should filter specific tokens when targets is "specific"', async () => {
-      const { CoverOverride } = await import(
-        '../../../scripts/rule-elements/operations/CoverOverride.js'
-      );
-      const tokenIds = ['target-1'];
-      const tokens = CoverOverride.getTargetTokens(mockSubjectToken, 'specific', null, tokenIds);
-
-      expect(tokens.length).toBe(1);
-      expect(tokens[0].id).toBe('target-1');
-    });
-
-    it('should apply range filter', async () => {
-      const { CoverOverride } = await import(
-        '../../../scripts/rule-elements/operations/CoverOverride.js'
-      );
-      global.canvas.grid.measureDistance = jest.fn().mockReturnValueOnce(5).mockReturnValueOnce(25);
-
-      const tokens = CoverOverride.getTargetTokens(mockSubjectToken, 'all', 20);
-
-      expect(tokens.length).toBe(1);
-    });
-  });
-
-  describe('checkDirectionalCover', () => {
-    it('should return true when attack is from blocked edge', async () => {
-      const { CoverOverride } = await import(
-        '../../../scripts/rule-elements/operations/CoverOverride.js'
-      );
-      const providerToken = { x: 100, y: 100 };
-      const receiverToken = { x: 100, y: 100 };
-      const attackerToken = { x: 100, y: 50 };
-
-      const isProtected = CoverOverride.checkDirectionalCover(
-        providerToken,
-        receiverToken,
-        attackerToken,
-        ['north'],
-      );
-
-      expect(isProtected).toBe(true);
-    });
-
-    it('should return false when attack is from unblocked edge', async () => {
-      const { CoverOverride } = await import(
-        '../../../scripts/rule-elements/operations/CoverOverride.js'
-      );
-      const providerToken = { x: 100, y: 100 };
-      const receiverToken = { x: 100, y: 100 };
-      const attackerToken = { x: 100, y: 150 };
-
-      const isProtected = CoverOverride.checkDirectionalCover(
-        providerToken,
-        receiverToken,
-        attackerToken,
-        ['north'],
-      );
-
-      expect(isProtected).toBe(false);
-    });
-  });
-});
 
 describe('Integration Tests', () => {
   describe('Hide Action Integration', () => {
